Move creating background into a seperate scene
