+++ application/frontend/src/game/scenes/MainMenu.ts
@@ -1,5 +1,6 @@
 import Phaser from "phaser"
 import type InventoryMenu from "./InventoryMenu"
+import type BaseScene from "./Base"
 
 export default class MainMenu extends Phaser.Scene {
   public minerTimers: { [key: string]: Phaser.Time.TimerEvent } = {}
@@ -16,42 +17,6 @@ export default class MainMenu extends Phaser.Scene {
   }
 
   editorCreate(): void {
-    // background_1
-    const background_1 = this.add.image(767, 431, "background_1")
-    background_1.scaleX = 1.5
-    background_1.scaleY = 1.5
-
-    const inventoryMenu = this.scene.get("InventoryMenu") as InventoryMenu
-    // iron_ore_block
-    const iron_ore_block = this.add.image(374, 247, "iron_ore_block")
-    iron_ore_block.scaleX = 7.8
-    iron_ore_block.scaleY = 7.8
-    iron_ore_block.setInteractive()
-    iron_ore_block.on("pointerdown", () => {
-      inventoryMenu.addToInventory("iron_ore", 1)
-      inventoryMenu.updateInventoryDisplay()
-    })
-
-    // copper_ore_block
-    const copper_ore_block = this.add.image(807, 312, "copper_ore_block")
-    copper_ore_block.scaleX = 7.8
-    copper_ore_block.scaleY = 7.8
-    copper_ore_block.setInteractive()
-    copper_ore_block.on("pointerdown", () => {
-      inventoryMenu.addToInventory("copper_ore", 1)
-      inventoryMenu.updateInventoryDisplay()
-    })
-
-    // rock_block
-    const rock_block = this.add.image(432, 435, "rock_block")
-    rock_block.scaleX = 7.8
-    rock_block.scaleY = 7.8
-    rock_block.setInteractive()
-    rock_block.on("pointerdown", () => {
-      inventoryMenu.addToInventory("rock", 1)
-      inventoryMenu.updateInventoryDisplay()
-    })
-
     // rectangle_1
     const rectangle_1 = this.add.rectangle(1195, 290, 128, 128)
     rectangle_1.scaleX = 1.447046084870738
@@ -118,6 +83,7 @@ export default class MainMenu extends Phaser.Scene {
   create() {
     this.scene.launch("CraftingMenu")
     this.scene.launch("InventoryMenu")
+    this.scene.launch("BaseScene")
     // this.scene.launch('RunningSmeltersScene');
     this.editorCreate()
